+++ packages/swapper/src/swappers/zrx/ZrxBuildQuoteTx/ZrxBuildQuoteTx.test.ts
@@ -5,7 +5,7 @@ import { ChainTypes, GetQuoteInput } from '@shapeshiftoss/types'
 import BigNumber from 'bignumber.js'
 import Web3 from 'web3'
 
-import { APPROVAL_GAS_LIMIT, DEFAULT_SLIPPAGE, MAX_SLIPPAGE } from '../utils/constants'
+import { APPROVAL_GAS_LIMIT, MAX_SLIPPAGE } from '../utils/constants'
 import { setupQuote } from '../utils/test-data/setupSwapQuote'
 import { zrxService } from '../utils/zrxService'
 import { ZrxBuildQuoteTx } from './ZrxBuildQuoteTx'
@@ -41,7 +41,6 @@ Web3.mockImplementation(() => ({
 
 const mockQuoteResponse = {
   allowanceContract: 'allowanceTargetAddress',
-  allowanceGrantRequired: true,
   buyAmount: undefined,
   buyAsset: {
     assetId: 'eip155:1/erc20:0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
@@ -76,8 +75,6 @@ const mockQuoteResponse = {
       gasPrice: undefined
     }
   },
-  guaranteedPrice: undefined,
-  priceImpact: undefined,
   rate: undefined,
   receiveAddress: '0xc770eefad204b5180df6a14ee197d99d808ee52d',
   sellAmount: '1000000000000000000',
@@ -105,9 +102,7 @@ const mockQuoteResponse = {
     tokenId: '0xc770eefad204b5180df6a14ee197d99d808ee52d'
   },
   sellAssetAccountId: '0',
-  slippage: DEFAULT_SLIPPAGE,
   sources: [{ name: '0x', proportion: '1' }],
-  statusCode: 0,
   success: true,
   txData: undefined
 }
@@ -296,7 +291,6 @@ describe('ZrxBuildQuoteTx', () => {
 
     expect(await ZrxBuildQuoteTx(deps, { input: quoteInput, wallet })).toEqual({
       success: false,
-      statusCode: 400,
       statusReason: 'Unknown Error',
       buyAsset: { ...mockQuoteResponse.buyAsset },
       sellAsset: { ...mockQuoteResponse.sellAsset }
@@ -310,7 +304,6 @@ describe('ZrxBuildQuoteTx', () => {
 
     expect(await ZrxBuildQuoteTx(deps, { input: quoteInput, wallet })).toEqual({
       success: false,
-      statusCode: 500,
       statusReason: 'Unknown Error',
       buyAsset: { ...mockQuoteResponse.buyAsset },
       sellAsset: { ...mockQuoteResponse.sellAsset }